fix: widen large fundamentals fields from long to decimal?
Live integration testing against EODHD surfaced a deserialization failure:
fundamentals fields like outstandingShares.*.shares, EBITDA, RevenueTTM,
SharesOutstanding/Float, EnterpriseValue, and crypto supply come back with decimals
or beyond Int64 range. Widen the long/long? fields in the Fundamental and
BulkFundamental trees to decimal? (genuine integer fields — exchange ticker counts,
EOD volume, unix timestamps — stay long).

Full integration suite now green against live data: 24/24.

+++ src/EODHD.CSharpApiClient/DataModel/Fundamental/ETFData.cs
@@ -131,7 +131,7 @@ namespace EODHD.CSharpApiClient.DataModel.Fundamental
         /// Gets or sets the number of holdings (count).
         /// </summary>
         [JsonPropertyName("Holdings_Count")]
-        public long? HoldingsCount { get; set; }
+        public decimal? HoldingsCount { get; set; }
 
         /// <summary>
         /// Gets or sets the top 10 holdings keyed by symbol.
